extract mousemove & pointerdown events listener on the editor
- the functions are better defined and should cover most cases for now.
It won't cover the cases where the sprites or objects are bigger than a
cell but that's a problem for future me.
The new way is also be easier to add new objects and will keep a
consistent structure between objects.

The `dispatch` function and the types should be handled in future
commits.