use reducer, actions & state to handle app state instead of `useState` in `MapContext`
I think it will be faster and more consistent to add actions this way,
instead of adding `useState` over and over.

Still need to think where to put extra information that won't be part of
the map itself but rather the config of the editor (currently selected
element for example).