fix textures not properly loading
when there are too many textures loading, ladder or spawn doesn't seem
to load and end up using another texture???

The problem is kind of random but doesn't appear when calling a function
to load the assets as opposed to using an exported object.

At the same time, I've removed the preview of the block using Graphics
in favor of Sprite. Might as well use the corresponding texture.