rename small_ramps to ramps & spawn to spawn_player
since there are only 2 types of ramps (normal & long), might as well
rename the normal ones.
There will be other kind of spawn, so might as well differentiate them
now.

Replace the type `PluralMapObjectName` in favor of `MapInformationKeys`.
I adapted the reducer to use a switch case instead. It does the same
thing, it's just easier to read.