user can draw by dragging the mouse when pointer is down
Also, the user now cannot place multiple instances of the same object on
the same coordinates.

For example, if a block has been placed to x:128, y:128; if the user
clicks or drags a block on it, no new block will be placed.

This behavior may be modified later to allow tiles to be placed on the
same coordinates due to layers/depth.
Currently, it does not make sense for our implemented objects.